Library

Documenting the user interface.

This page is currently broken and won't show most of the functions for some reason, any suggestions would be appreciated!

OceanBioME.jl

OceanBioME.error_on_neg! — Method

error_on_neg(sim; params = (exclude=(), ))

Throws an error if any tracers in sim.model are negative. Use like:

simulation.callbacks[:neg] = Callback(error_on_neg!)

Tracers to exclude can be set in the parameters.

OceanBioME.warn_on_neg! — Method

warn_on_neg(sim; params = (exclude=(), ))

Raises a warning if any tracers in sim.model are negative. Use like:

simulation.callbacks[:neg] = Callback(warn_on_neg!)

Tracers to exclude can be set in the parameters.

OceanBioME.zero_negative_tracers! — Method

zero_negative_tracers!(sim; params = (exclude=(), warn=false))

Sets any tracers in sim.model which are negative to zero. Use like:

simulation.callbacks[:neg] = Callback(zero_negative_tracers!)

This is NOT a reccomended method to preserve positivity as it strongly does not conserve tracers.

Tracers to exclude can be set in the parameters and if params.warn is set to true a warning will be displayed when negative values are modified.

Light

OceanBioME.Light — Module

Light attenuation by chlorophyll as described by P. Karleskind, M. Lévy, L. Memery (2011) (implimented as twoBand) and André Morel (1988).

These should be used by setting up a callback like: simulation.callbacks[:update_par] = Callback(Light.update_2λ!, IterationInterval(1), merge(params, (surface_PAR=surface_PAR,)), TimeStepCallsite())

Boundary Conditions

OceanBioME.Boundaries — Module

Boundary conditions for air/sea and sediment flux.

Currently implimented:

gasexchange (R. Wanninkhof (1992))

Generic air sea flux model describted by Wanninkhof, 1992 but only setup for CO₂ and O₂

Forces the DIC and ocygen fields, and requires temp (in centigrade) and salinity, plus current DIC and ALK concentration

Sediments

Soetaert (Karline Soetaert, Jack J Middelburg, Peter M J Herman, Kerst Buis (2000))

simple (integrated) sediment model described by Soetaert, Middelburg, Herman and Buis, 2000

where organic matter (D and DD) that sinks to the bottom is stored and decays into NO₃ and NH₄, and takes up O₂ in the process.

Extended to attribute the corresponding release of DIC

Forced by O₂, NO₃, NH₄ and particle concentration in bottom cell

OceanBioME.Boundaries.Sediments.Soetaert — Module

Soetaert

Soetaert et al. 2000 vertically integrated sediment diagenesis model.

XXX Currently does not work properly and will result in nitrogen leak!

Notes

In the paper p functions are determined from fitting data and are not closed (e.g. if anything is zero they go to NaN or Inf)

Additionally theres a lot of log and exp of dimensional quantities

Units may also be wrong

OceanBioME.Boundaries.Sediments.Soetaert.setup — Method

Boundaries.Sediments.Soetaert.setup(grid, POM_fields::NamedTuple; 
                                                            POM_w = (DD = 200/day, D = 3.47e-5), 
                                                            parameters=defaults, 
                                                            Nᵣᵣᵢ=24.0, 
                                                            Nᵣᵢ=85.0, 
                                                            f_ref=0.1, 
                                                            f_fast=0.74,
                                                            f_slow=0.26, 
                                                            carbonates=true)

Returns the boundary conditions, forcing, auxiliary fields, and parameters for the Karline Soetaert, Jack J Middelburg, Peter M J Herman, Kerst Buis (2000) vertically integrated sediment diagenesis model.

Arguments

grid: Oceananigans grid.

POM_fields: NamedTuple of Oceananigans fields which contribute organic matter to the sediment

Keyword arguments

POM_w: NamedTuple of the sinking speed (at the bottom of the domain) for each sinking tracer

Nᵣᵣᵢ and Nᵣᵢ: initial values for the fast and slow degrading sediment components

f_ref, f_fast, and f_slow: fraction of sedimened organic matter which is: refractory, fast degrading and slow degrading. f_ref is a fraction of the total depositional flux where as the others are a fraction of the remaining flux

carbonates: as a temporary measure this is used to determine if a DIC boundary condition should be provided

Notes

Fast/slow/refractory fractions from B.~P. Boudreau, B.~R. Ruddick (1991) and T K Tromp, P Van Cappellen, R M Key (1995) as used by Karline Soetaert, Jack J Middelburg, Peter M J Herman, Kerst Buis (2000) and initial values from steadyt state average deposition after 1 year without sediment in sub polar region (approx 0.01 mmolN/m³ for DD and 0 for D).

Box Model

OceanBioME.BoxModels — Module

Integrate biogeochemical models on a single point

OceanBioME.BoxModels.SaveBoxModel — Type

SaveBoxModel(filepath::FP)

Construct object to save box model outputs at filepath.

Arguments: filepath - path to save results to

Oceananigans.Fields.set! — Method

set!(model::BoxModel; kwargs...)

Set the values for a BoxModel

Arguments: model - the model to set the arguments for Keyword Arguments: variable and value pairs to set

Oceananigans.Simulations.run! — Method

run!(model::BoxModel; feedback_interval = 1000, save_interval = Inf, save = nothing)

Run a box model

Arguments: model - the BoxModel to solve

Keyword Arguments

feedback_interval: how often (number of iterations) to display progress

save_interval: how often (number of iterations) to save output

save: SaveBoxModel object to specify how to save output

TODO: should abstract out to simulation like Oceananians to add e.g. callbacks