Library

Documenting the user interface.

OceanBioME.jl

OceanBioME.OceanBioME — Module

A fast and flexible modelling environment for modelling the coupled interactions between ocean biogeochemistry, carbonate chemistry, and physics.

OceanBioME.ScaleNegativeTracers — Method

ScaleNegativeTracers(; tracers, scalefactors = ones(length(tracers)), warn = false)

Returns a callback that scales tracers so that none are negative. Use like:

negativity_protection! = ScaleNegativeTracers(; model, tracers = (:P, :Z, :N))
simulation.callbacks[:neg] = Callback(negativity_protection!; callsite = UpdateStateCallsite())

This method is better, though still imperfect, method to prevent numerical errors that lead to negative tracer values compared to zero_negative_tracers!. Please see discussion in github.

Future plansj include implement a positivity=preserving timestepping scheme as the ideal alternative.

OceanBioME.error_on_neg! — Method

error_on_neg(sim; params = (exclude=(), ))

Throws an error if any tracers in sim.model are negative. Use like:

simulation.callbacks[:neg] = Callback(error_on_neg!)

Tracers to exclude can be set in the parameters.

OceanBioME.remove_NaN_tendencies! — Method

remove_NaN_tendencies!(model)

Zeros any NaN value tendencies as a final protection against negative tracer run away.

OceanBioME.warn_on_neg! — Method

warn_on_neg(sim; params = (exclude=(), ))

Raises a warning if any tracers in sim.model are negative. Use like:

simulation.callbacks[:neg] = Callback(warn_on_neg!)

Tracers to exclude can be set in the parameters.

OceanBioME.zero_negative_tracers! — Method

zero_negative_tracers!(sim; params = (exclude=(), ))

Sets any tracers in sim.model which are negative to zero. Use like:

simulation.callbacks[:neg] = Callback(zero_negative_tracers!)

Tracers to exclude can be set in the parameters.

Tracer conservation

This method is not recommended as a way to preserve positivity of tracers since it does not conserve the total tracer.

Biogeochemical Models

Nutrient Phytoplankton Zooplankton Detritus

OceanBioME.NPZDModel — Module

Nutrient-Phytoplankton-Zooplankton-Detritus model of Angela M. Kuhn, Katja Fennel, Jann Paul Mattern (2015)

Tracers

Nutrients: N (mmol N/m³)

Phytoplankton: P (mmol N/m³)

Zooplankton: Z (mmol N/m³)

Detritus: D (mmol N/m³)

Required submodels

Photosynthetically available radiation: PAR (W/m²)

The Lodyc-DAMTP Ocean Biogeochemical Simulation Tools for Ecosystem and Resources (LOBSTER)

OceanBioME.LOBSTERModel — Module

The Lodyc-DAMTP Ocean Biogeochemical Simulation Tools for Ecosystem and Resources (LOBSTER) model.

Tracers

Nitrates: NO₃ (mmol N/m³)

Ammonia: NH₄ (mmol N/m³)

Phytoplankton: P (mmol N/m³)

Zooplankton: Z (mmol N/m³)

Small (slow sinking) particulate organic matter: sPOM (mmol N/m³)

Large (fast sinking) particulate organic matter: bPOM (mmol N/m³)

Disolved organic matter: DOM (mmol N/m³)

Optional tracers

Carbonate chemistry

Disolved inorganic carbon: DIC (mmol C/m³)

Alkalinity: Alk (meq/m³)

Oxygen chemistry

Oxygen: O₂ (mmol O₂/m³)

Variable redfield

Small (slow sinking) particulate organic matter carbon content: sPOC (mmol C/m³)

Large (fast sinking) particulate organic matter carbon content: bPOC (mmol C/m³)

Disolved organic matter carbon content: DOC (mmol C/m³)

When this option is enabled then the usual sPOM and bPOM change to sPON and bPON as they explicitly represent the nitrogen contained in the particulate matter

Required submodels

Photosynthetically available radiation: PAR (W/m²)

For optional tracers:

Temperature: T (ᵒC)

Salinity: S (‰)

OceanBioME.LOBSTERModel.LOBSTER — Type

LOBSTER(; grid,
          phytoplankton_preference::FT = 0.5,
          maximum_grazing_rate::FT = 9.26e-6, # 1/s
          grazing_half_saturation::FT = 1.0, # mmol N/m³
          light_half_saturation::FT = 33.0, # W/m² (?)
          nitrate_ammonia_inhibition::FT = 3.0,
          nitrate_half_saturation::FT = 0.7, # mmol N/m³
          ammonia_half_saturation::FT = 0.001, # mmol N/m³
          maximum_phytoplankton_growthrate::FT = 1.21e-5, # 1/s
          zooplankton_assimilation_fraction::FT = 0.7,
          zooplankton_mortality::FT = 2.31e-6, # 1/s/mmol N/m³
          zooplankton_excretion_rate::FT = 5.8e-7, # 1/s
          phytoplankton_mortality::FT = 5.8e-7, # 1/s
          small_detritus_remineralisation_rate::FT = 5.88e-7, # 1/s
          large_detritus_remineralisation_rate::FT = 5.88e-7, # 1/s
          phytoplankton_exudation_fraction::FT = 0.05,
          nitrifcaiton_rate::FT = 5.8e-7, # 1/s
          ammonia_fraction_of_exudate::FT = 0.75, 
          ammonia_fraction_of_excriment::FT = 0.5,
          ammonia_fraction_of_detritus::FT = 0.0,
          phytoplankton_redfield::FT = 6.56, # mol C/mol N
          organic_redfield::FT = 6.56, # mol C/mol N
          phytoplankton_chlorophyll_ratio::FT = 1.31, # g Chl/mol N
          organic_carbon_calcate_ratio::FT = 0.1, # mol CaCO₃/mol C
          respiraiton_oxygen_nitrogen_ratio::FT = 10.75, # mol O/molN
          nitrifcation_oxygen_nitrogen_ratio::FT = 2.0, # mol O/molN
          slow_sinking_mortality_fraction::FT = 0.5, 
          fast_sinking_mortality_fraction::FT = 0.5,
          disolved_organic_breakdown_rate::FT = 3.86e-7, # 1/s
          zooplankton_calcite_dissolution::FT = 0.3,

          surface_phytosynthetically_active_radiation::SPAR = (x, y, t) -> 100*max(0.0, cos(t*π/(12hours))),

          light_attenuation_model = TwoBandPhotosyntheticallyActiveRatiation(; grid),
          sediment_model::S = nothing,

          carbonates::Bool = false,
          oxygen::Bool = false,
          variable_redfield = false,

          sinking_speed = (sPOM = 3.47e-5, bPOM = 200/day),
          open_bottom::Bool = true,

          particles::P = nothing)

Construct an instance of the LOBSTER (LOBSTER) biogeochemical model.

Keywork Arguments

grid: (required) the geometry to build the model on, required to calculate sinking

phytoplankton_preference, ..., disolved_organic_breakdown_rate: LOBSTER parameter values

surface_phytosynthetically_active_radiation: funciton (or array in the future) for the photosynthetically available radiation at the surface, should be shape f(x, y, t)

light_attenuation_model: light attenuation model which integrated the attenuation of available light

sediment_model: slot for AbstractSediment

carbonates, oxygen, and variable_redfield: include models for carbonate chemistry and/or oxygen chemistry and/or variable redfield ratio disolved and particulate organic matter

sinking_speed: named tuple of constant sinking, of fields (i.e. ZFaceField(...)) for any tracers which sink (convention is that a sinking speed is positive, but a field will need to follow the usual down being negative)

open_bottom: should the sinking velocity be smoothly brought to zero at the bottom to prevent the tracers leaving the domain

particles: slot for BiogeochemicalParticles

Sugar kelp (Saccharina latissima)

OceanBioME.SLatissimaModel — Module

Sugar kelp model of Ole Jacob Broch, Dag Slagstad (2012) and updated by Ole Jacob Broch, Ingrid Helene Ellingsen, Silje Forbord, Xinxin Wang, Zsolt Volent, Morten Omholt Alver, Aleksander Handå, Kjersti Andresen, Dag Slagstad, Kjell Inge Reitan, Yngvar Olsen, Jorunn Skjermo (2013), Julia Fossberg, Silje Forbord, Ole Jacob Broch, Arne M. Malzahn, Henrice Jansen, Aleksander Handå, Henny Førde, Maria Bergvik, Anne Lise Fleddum, Jorunn Skjermo, Yngvar Olsen (2018), and Ole Jacob Broch, Morten Omholt Alver, Trine Bekkby, Hege Gundersen, Silje Forbord, Aleksander Handå, Jorunn Skjermo, Kasper Hancke (2019).

Prognostic properties

Area: A (dm²)

Nitrogen reserve: N (gN/gSW)

Carbon reserve: C (gC/gSW)

Tracer dependencies

Nitrates: NO₃ (mmol N/m³)

Photosynthetically available radiation: PAR (einstein/m²/day)

Optionally:

Ammonia: NH₄ (mmol N/m³)

OceanBioME.SLatissimaModel.SLatissima — Type

SLatissima(; architecture :: AR = CPU()
             growth_rate_adjustement :: FT = 4.5
             photosynthetic_efficiency :: FT = 4.15e-5 * 24 * 10^6 / (24 * 60 * 60)
             minimum_carbon_reserve :: FT = 0.01
             structural_carbon :: FT = 0.2
             exudation :: FT = 0.5
             erosion :: FT = 0.22
             saturation_irradiance :: FT = 90 * day/ (10 ^ 6)
             structural_dry_weight_per_area :: FT = 0.5
             structural_dry_to_wet_weight :: FT = 0.0785
             carbon_reserve_per_carbon :: FT = 2.1213
             nitrogen_reserve_per_nitrogen :: FT = 2.72
             minimum_nitrogen_reserve :: FT = 0.0126
             maximum_nitrogen_reserve :: FT = 0.0216
             growth_adjustement_2 :: FT = 0.039 / (2 * (1 - minimum_nitrogen_reserve / maximum_nitrogen_reserve))
             growth_adjustement_1 :: FT = 0.18 / (2 * (1 - minimum_nitrogen_reserve / maximum_nitrogen_reserve)) - growth_adjustement_2
             maximum_specific_growth_rate :: FT = 0.18
             structural_nitrogen :: FT = 0.0146
             photosynthesis_at_ref_temp_1 :: FT = 1.22e-3 * 24
             photosynthesis_at_ref_temp_2 :: FT = 1.3e-3 * 24
             photosynthesis_ref_temp_1 :: FT = 285.0
             photosynthesis_ref_temp_2 :: FT = 288.0
             photoperiod_1 :: FT = 0.85
             photoperiod_2 :: FT = 0.3
             respiration_at_ref_temp_1 :: FT = 2.785e-4 * 24
             respiration_at_ref_temp_2 :: FT = 5.429e-4 * 24
             respiration_ref_temp_1 :: FT = 285.0
             respiration_ref_temp_2 :: FT = 290.0
             photosynthesis_arrhenius_temp :: FT = (1 / photosynthesis_ref_temp_1 - 1 / photosynthesis_ref_temp_2) ^ -1 * log(photosynthesis_at_ref_temp_2 / photosynthesis_at_ref_temp_1)
             photosynthesis_low_temp :: FT = 271.0
             photosynthesis_high_temp :: FT = 296.0
             photosynthesis_high_arrhenius_temp :: FT = 1414.87
             photosynthesis_low_arrhenius_temp :: FT = 4547.89
             respiration_arrhenius_temp :: FT = (1 / respiration_ref_temp_1 - 1 / respiration_ref_temp_2) ^ -1 * log(respiration_at_ref_temp_2 / respiration_at_ref_temp_1)
             current_speed_for_0p65_uptake :: FT = 0.03
             nitrate_half_saturation :: FT = 4.0
             ammonia_half_saturation :: FT = 1.3
             maximum_nitrate_uptake :: FT = 10 * structural_dry_weight_per_area * 24 * 14 / (10^6)
             maximum_ammonia_uptake :: FT = 12 * structural_dry_weight_per_area * 24 * 14 / (10^6)
             current_1 :: FT = 0.72
             current_2 :: FT = 0.28
             current_3 :: FT = 0.045
             respiration_reference_A :: FT = 1.11e-4 * 24
             respiration_reference_B :: FT = 5.57e-5 * 24
             exudation_redfield_ratio :: FT = Inf

             pescribed_velocity :: U = 0.1
             pescribed_temperature :: T = nothing
             pescribed_salinity :: S = nothing

             #position
             x :: P = arch_array(architecture, [0.0])
             y :: P = arch_array(architecture, zeros(Float64, length(x)))
             z :: P = arch_array(architecture, zeros(Float64, length(x)))

             #properties
             A :: P = arch_array(architecture, ones(Float64, length(x)) * 30)
             N :: P = arch_array(architecture, ones(Float64, length(x)) * 0.01)
             C :: P = arch_array(architecture, ones(Float64, length(x)) * 0.1)

             #feedback
             nitrate_uptake :: P = arch_array(architecture, zeros(Float64, length(x)))
             ammonia_uptake :: P = arch_array(architecture, zeros(Float64, length(x)))
             primary_production :: P = arch_array(architecture, zeros(Float64, length(x)))
             frond_exudation :: P = arch_array(architecture, zeros(Float64, length(x)))
             nitrogen_erosion :: P = arch_array(architecture, zeros(Float64, length(x)))
             carbon_erosion :: P = arch_array(architecture, zeros(Float64, length(x)))

             custom_dynamics :: F = no_dynamics

             scalefactor :: FT = 1.0
             latitude :: FT = 57.5)

Keywork Arguments

architecture: the architecture to adapt arrays to

growth_rate_adjustement, ..., exudation_redfield_ratio: parameter values

pescribed_velocity, pescribed_temperature and pescribed_salinity: functions for the relative velocity, temperature and salinity in the form f(x, y, z, t)

x,y and z: positions of the particles

A, N, and C: area, nitrogen, and carbon reserves

nitrate_uptake ... carbon_erosion: diagnostic values coupled to tracer fields

custom_dynamics: place to add any function of form f!(particles, model, bgc, Δt)

scalefactor: scalar scaling for tracer coupling

latitude: model latitude for seasonal growth modulation

Light Attenuation Models

OceanBioME.Light — Module

Light attenuation by chlorophyll as described by P. Karleskind, M. Lévy, L. Memery (2011) (implemented as twoBand) and André Morel (1988).

OceanBioME.Light.TwoBandPhotosyntheticallyActiveRatiation — Method

TwoBandPhotosyntheticallyActiveRatiation(; grid, 
                                           water_red_attenuation::FT = 0.225, # 1/m
                                           water_blue_attenuation::FT = 0.0232, # 1/m
                                           chlorophyll_red_attenuation::FT = 0.037, # 1/(m * (mgChl/m³) ^ eʳ)
                                           chlorophyll_blue_attenuation::FT = 0.074, # 1/(m * (mgChl/m³) ^ eᵇ)
                                           chlorophyll_red_exponent::FT = 0.629,
                                           chlorophyll_blue_exponent::FT = 0.674,
                                           pigment_ratio::FT = 0.7,
                                           phytoplankton_chlorophyll_ratio::FT = 1.31,
                                           surface_PAR::SPAR = (x, y, t) -> 100 * max(0.0, cos(t * π / 12hours)))

Keywork Arguments

grid: grid for building the model on

water_red_attenuation, ..., phytoplankton_chlorophyll_ratio: parameter values

surface_PAR: function (or array in the future) for the photosynthetically available radiaiton at the surface, should be shape f(x, y, t)

Boundary Conditions

OceanBioME.Boundaries — Module

Boundary conditions for air/sea and sediment flux.

Currently implemented:

gasexchange (R. Wanninkhof (1992))

Generic air-sea flux model described by Wanninkhof, 1992 but only setup for CO₂ and O₂.

Forces the DIC and oxygen fields, and requires temp (in centigrade) and salinity, plus current DIC and ALK concentration.

Sediments

Soetaert (Karline Soetaert, Jack J Middelburg, Peter M J Herman, Kerst Buis (2000))

simple (integrated) sediment model described by Soetaert, Middelburg, Herman and Buis, 2000 where organic matter (D and DD) that sinks to the bottom is stored and decays into NO₃ and NH₄, and takes up O₂ in the process.

Extended to attribute the corresponding release of DIC.

Forced by O₂, NO₃, NH₄ and particle concentration in bottom cell.

OceanBioME.Boundaries.Sediments.SimpleMultiG — Type

struct SimpleMultiG

Hold the parameters and fields for a simple "multi G" single-layer sediment model. Based on the Level 3 model described by Soetaert et al. 2000; doi:10.1016/S0012-8252(00)00004-0.

OceanBioME.Boundaries.Sediments.SimpleMultiG — Method

SimpleMultiG(grid; 
             fast_decay_rate::FT = 2/day,
             slow_decay_rate::FT = 0.2/day,
             fast_redfield::FT = 0.1509,
             slow_redfield::FT = 0.13,
             fast_fraction::FT = 0.74,
             slow_fraction::FT = 0.26,
             refactory_fraction::FT = 0.1,
             nitrate_oxidation_params::P1 = (A = - 1.9785, B = 0.2261, C = -0.0615, D = -0.0289, E = - 0.36109, F = - 0.0232),
             denitrifcaiton_params::P2 = (A = - 3.0790, B = 1.7509, C = 0.0593, D = - 0.1923, E = 0.0604, F = 0.0662),
             anoxic_params::P3 = (A = - 3.9476, B = 2.6269, C = - 0.2426, D = -1.3349, E = 0.1826, F = - 0.0143),
             depth = abs(znode(1, grid, Face())),
             solid_dep_params::P4 = (A = 0.233, B = 0.336, C = 982, D = - 1.548, depth = depth))

Returns a single-layer "multi G" sediment model (SimpleMultiG) on grid where parameters can be optionally specified.

The model is a single layer (i.e. does not include porous diffusion) model with three classes of sediment organic matter which decay at three different rates (fast, slow, refactory). The nitrifcation/denitrifcation/anoxic mineralisation fractions default to the parameterisation of Soetaert et al. 2000; doi:10.1016/S0012-8252(00)00004-0.

This model has not yet been validated or compared to observational data. The variety of degridation processes is likely to be strongly dependent on oxygen availability (see https://bg.copernicus.org/articles/6/1273/2009/bg-6-1273-2009.pdf) so it will therefore be important to also thoroughly validate the oxygen model (also currently limited).

Example

using OceanBioME, Oceananigans, OceanBioME.Sediments

grid = RectilinearGrid(size=(3, 3, 30), extent=(10, 10, 200))

sediment_model = SimpleMultiG(grid)

Box Model

OceanBioME.BoxModels — Module

Integrate biogeochemical models on a single point

OceanBioME.BoxModels.SaveBoxModel — Type

SaveBoxModel(filepath::FP)

Construct object to save box model outputs at filepath.

Arguments: filepath - path to save results to

Oceananigans.Fields.set! — Method

set!(model::BoxModel; kwargs...)

Set the values for a BoxModel

Arguments: model - the model to set the arguments for Keyword Arguments: variable and value pairs to set

Oceananigans.Simulations.run! — Method

run!(model::BoxModel; feedback_interval = 1000, save_interval = Inf, save = nothing)

Run a box model

Arguments: model - the BoxModel to solve

Keyword Arguments

feedback_interval: how often (number of iterations) to display progress

save_interval: how often (number of iterations) to save output

save: SaveBoxModel object to specify how to save output

TODO: should abstract out to simulation like Oceananians to add e.g. callbacks